move use_fake_lammps to utils
